Upgrade to systemd-stable v208
Additional changes in ptest code since now
we have directories and not only bunch of files
under test/ dir so a simple install does not
work anymore we have to cp the files


Conflicts:
	meta/recipes-core/systemd/systemd/systemd-pam-fix-fallocate.patch
	meta/recipes-core/systemd/systemd_208.bb

+++ meta/recipes-core/systemd/systemd_208.bb
@@ -17,14 +17,12 @@ SECTION = "base/shell"
 
 inherit gtk-doc useradd pkgconfig autotools perlnative update-rc.d update-alternatives qemu systemd ptest
 
-SRC_URI = "http://www.freedesktop.org/software/systemd/systemd-${PV}.tar.xz \
-           file://0001-Use-bin-mkdir-instead-of-host-mkdir-path.patch \
+SRCREV = "255eb046a7bcb90e60a3a54302bc1250c1aed26a"
+
+PV = "208+git${SRCPV}"
+
+SRC_URI = "git://anongit.freedesktop.org/systemd/systemd-stable;branch=v208-stable;protocol=git \
            file://binfmt-install.patch \
-           file://journald-add-missing-error-check.patch \
-           file://journald-fix-minor-memory-leak.patch \
-           file://journal-when-appending-to-journal-file-allocate-larg.patch \
-           file://journal-file-protect-against-alloca-0.patch \
-           file://journal-Add-missing-byte-order-conversions.patch \
            file://touchscreen.rules \
            ${UCLIBCPATCHES} \
            file://00-create-volatile.conf \
@@ -32,8 +30,8 @@ SRC_URI = "http://www.freedesktop.org/software/systemd/systemd-${PV}.tar.xz \
            file://var-run.conf \
            file://run-ptest \
           "
-SRC_URI[md5sum] = "df64550d92afbffb4f67a434193ee165"
-SRC_URI[sha256sum] = "aa64fa864466fd5727005c55d61c092828b94b4f857272c0b503695022146390"
+
+S = "${WORKDIR}/git"
 
 UCLIBCPATCHES = ""
 UCLIBCPATCHES_libc-uclibc = "file://systemd-pam-configure-check-uclibc.patch \
@@ -77,7 +75,6 @@ EXTRA_OECONF = " --with-rootprefix=${rootprefix} \
                  --disable-manpages \
                  --disable-coredump \
                  --disable-introspection \
-                 --disable-tcpwrap \
                  --enable-split-usr \
                  --without-python \
                  --with-sysvrcnd-path=${sysconfdir} \
@@ -100,8 +97,8 @@ do_install() {
 	rm ${D}${systemd_unitdir}/system/serial-getty* -f
 
 	# Provide support for initramfs
-	ln -s ${rootlibexecdir}/systemd/systemd ${D}/init
-	ln -s ${rootlibexecdir}/systemd/systemd-udevd ${D}/${base_sbindir}/udevd
+	[ ! -e ${D}/init ] && ln -s ${rootlibexecdir}/systemd/systemd ${D}/init
+	[ ! -e ${D}/${base_sbindir}/udevd ] && ln -s ${rootlibexecdir}/systemd/systemd-udevd ${D}/${base_sbindir}/udevd
 
 	# Create dir for journal
 	install -d ${D}${localstatedir}/log/journal
@@ -131,7 +128,7 @@ do_install() {
 
 do_install_ptest () {
        install -d ${D}${PTEST_PATH}/test
-       install ${S}/test/* ${D}${PTEST_PATH}/test
+       cp -rf ${S}/test/* ${D}${PTEST_PATH}/test
        install -m 0755  ${B}/test-udev ${D}${PTEST_PATH}/
        install -d ${D}${PTEST_PATH}/build-aux
        cp -rf ${B}/rules ${D}${PTEST_PATH}/
@@ -161,6 +158,12 @@ FILES_${PN}-analyze = "${bindir}/systemd-analyze"
 
 FILES_${PN}-initramfs = "/init"
 RDEPENDS_${PN}-initramfs = "${PN}"
+
+# The test cases need perl and bash to run correctly.
+RDEPENDS_${PN}-ptest += "perl bash"
+FILES_${PN}-ptest += "${libdir}/udev/rules.d"
+
+FILES_${PN}-dbg += "${libdir}/systemd/ptest/.debug"
 
 FILES_${PN}-gui = "${bindir}/systemadm"
